Added: All the questions for the questionnaire
Hi,
I added all the questions except the BMI calculator. In the commented text, you can some discussion points. I'd like the feedback of the aesthetics of the form and some of the options/answer boxes.

P.S I added our logo as a page icon for the start page (or the tab)!