Solved: The issue with the pagelink from Predictive to Prescriptive page
On the page 3_Predictive_Analytics, I commented the part where the default code from Luis was to load the model. This caused an error since we don't have a trained model yet to load into the code. When we do, we can uncomment this part! (OBS This is my very qualified killgissning)

As for the page links/page redirections: there was just a minor fault in the path. See the code, I kept both the old code (commented, please remove this after you checked)  and the new code.